Title: 체질량지수, Time: 45ms, Memory: 16MB, Status: Passed - Codetree

+++ 260203/체질량지수/bmi.py
@@ -1,6 +1,6 @@
 h, w = map(int, input().split())
 
-b = (10000 * w) / (h * h)
+b = int((10000 * w) / (h * h))
 
 print(b)
